requirements.txt
appdirs==1.4.4
autopep8==1.5.7
bcrypt==3.2.0
Bcrypt-Flask==1.0.1
certifi==2020.12.5
cffi==1.14.5
click==8.0.1
colorama==0.4.4
cycler==0.10.0
distlib==0.3.2
dnspython==2.1.0
filelock==3.0.12
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.1
Flask-PyMongo==2.3.0
h5py==3.2.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.20.3
opencv-contrib-python==4.5.2.54
pandas==1.2.4
Pillow==8.2.0
pipenv==2021.5.29
pybase64==1.1.4
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
toml==0.10.2
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tqdm==4.61.1
typing-extensions==3.10.0.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
Werkzeug==2.0.1
